Tiny fixes and refactoring
